AIMS scripts improvements - no bash
 * cleaning python script calls from bash
 * update README
 * remove various symbolic links
 * merge data_validation unittests into main script
 * add dotenv
