--- ANMN/NRS_AIMS/REALTIME/env
@@ -1,1 +1,0 @@
-../../../env
