--- ANMN/NRS_AIMS/REALTIME/lib
@@ -1,1 +1,0 @@
-../../../lib/
