+++ FAIMMS/REALTIME/README.md
@@ -13,18 +13,15 @@ data comes from the following sites :
 * Heron Island
 * Davies Reef
 
-## local testing
-
-Open config.txt to change the paths of :
-* data_wip_path                    :location of the FAIMMS scripts output (log + data)
-
-For testing on local machine, please edit $DATA_SERVICE_DIR/env
-
-Type in your shell ```./FAIMMS.sh```
-
+```bash
+./faimms.py
+```
 ## System Requirements
-Internet connection
-Tested Ubuntu 9.10, MINT17
+Python:
+see requirements.txt
+
+Operating System Support:
+POBOX
 
 ## Contact Support
 for support contact:
