--- SOOP/SOOP_TRV/lib
@@ -1,1 +1,0 @@
-../../lib/
