+++ cron.d/FAIMMS_RSS
@@ -2,4 +2,4 @@ SHELL=/bin/bash
 
 MAILTO=laurent.besnard@utas.edu.au
 
-0 * * * * projectofficer FAIMMS/REALTIME/FAIMMS.sh
+0 * * * * projectofficer FAIMMS/REALTIME/faimms.py
