+++ lib/netcdf/imos_env
@@ -1,0 +1,6 @@
+ACKNOWLEDGEMENT='Any users of IMOS data are required to clearly acknowledge the source of the material derived from IMOS in the format: "Data was sourced from the Integrated Marine Observing System (IMOS) - IMOS is a national collaborative research infrastructure, supported by the Australian Government." If relevant, also credit other organisations involved in collection of this particular datastream (as listed in "credit" in the metadata record).'
+CONVENTIONS='CF-1.6,IMOS-1.3'
+DATA_CENTRE='eMarine Information Infrastructure (eMII)'
+DATA_CENTRE_EMAIL='info@emii.org.au'
+DISTRIBUTION_STATEMENT='Data may be re-used, provided that related metadata explaining the data has been reviewed by the user, and the data is appropriately acknowledged. Data, products and services from IMOS are provided "as is" without any warranty as to fitness for a particular purpose.'
+PROJECT='Integrated Marine Observing System (IMOS)'
